+++ sphinx_collections/drivers/copy_file.py
@@ -1,5 +1,5 @@
 import os
-from shutil import copyfile
+from shutil import copy2
 
 from sphinx_collections.drivers import Driver
 
@@ -8,14 +8,20 @@ class CopyFileDriver(Driver):
     def run(self):
         self.info("Copy file...")
 
-        if not os.path.exists(self.config["source"]):
-            self.error("Source {} does not exist".format(self.config["source"]))
+        source = self.config["source"] if os.path.exists(self.config["source"]) else self.get_source_path()
+
+        if not os.path.exists(source):
+            self.error(f"Source {source} does not exist")
             return
 
+        target = self.config["target"]
+        self.create_target_dir(target)
+
         try:
-            copyfile(self.config["source"], self.config["target"])
-        except OSError as e:
-            self.error("Problems during copying file.", e)
+            self.info(f"cp {source} {target}")
+            copy2(source, target)
+        except Exception as e:
+            self.error(f"copy file from {source} to {target} failed: {e}")
 
     def clean(self):
         try: