+++ sphinx_collections/drivers/copy_folder.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 from shutil import copytree, ignore_patterns, rmtree
 
 from sphinx_collections.drivers import Driver
@@ -8,14 +9,17 @@ class CopyFolderDriver(Driver):
     def run(self):
         self.info("Copy folder...")
 
-        if not os.path.exists(self.config["source"]):
-            self.error("Source {} does not exist".format(self.config["source"]))
+        source = self.config["source"] if os.path.exists(self.config["source"]) else self.get_source_path()
+
+        if not os.path.exists(source):
+            self.error(f"Source {source} does not exist")
             return
 
+        target = self.config["target"].rstrip("/")
+        self.create_target_dir(str(Path(target).parent))
+
         try:
-            copytree(
-                self.config["source"], self.config["target"], ignore=ignore_patterns(*self.config.get("ignore", []))
-            )
+            copytree(source, target, ignore=ignore_patterns(*self.config.get("ignore", [])))
         except OSError as e:
             self.error("Problems during copying folder.", e)