+++ tests/test_config_symlink.py
@@ -5,7 +5,14 @@ import pytest
 
 @pytest.mark.parametrize(
     "test_app",
-    [{"buildername": "html", "srcdir": "doc_test/config_symlink", "datadir": "doc_test/config_symlink_data"}],
+    [
+        {
+            "buildername": "html",
+            "srcdir": "doc_test/config_symlink",
+            "datadir": "doc_test/config_symlink_data",
+            "confoverrides": {"collections_final_clean": False},
+        }
+    ],
     indirect=True,
 )
 def test_config_symlink(test_app):
@@ -14,6 +21,9 @@ def test_config_symlink(test_app):
     html = Path(app.outdir, "index.html").read_text()
 
     assert '<link rel="next" title="Awesome" href="_collections/my_data/awesome.html" />' in html
+
+    target = Path(app.confdir, "_collections/my_data")
+    assert target.is_symlink()
 
     html = Path(app.outdir, "_collections/my_data/awesome.html").read_text()